+++ mkfs.c
@@ -388,7 +388,7 @@ int main(int ac, char **av)
 					fprintf(stderr, "File system size "
 						"%llu bytes is too small, "
 						"256M is required at least\n",
-						block_count);
+						(unsigned long long)block_count);
 					exit(1);
 				}
 				zero_end = 0;
